06-ejemplos-graphql/01-flask-graphql/requirements.txt
flask>=3.0
strawberry-graphql[flask]>=0.260
pydantic>=2.0
